requirements.txt
anyio==4.0.0
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.2
future==1.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.0
opencv-python==4.8.1.78
pillow==10.2.0
python-dotenv==1.0.1
python-telegram-bot==20.6
sniffio==1.3.0
tinydb==4.8.0
Werkzeug==3.0.1
